Converts a slide deck to an HTML slideshow

Converts a slide deck to an HTML slideshow

Usage

make_reveal(
  repo,
  slides_md = file.path(repo, "slides.md"),
  output = file.path(normalizePath(repo), "slides.html"),
  extra_flags = character(),
  verbose = FALSE,
  open = TRUE
)

Arguments

repo

Path to the git repository corresponding to a lesson that uses Carpentries Workbench.

slides_md

Optional output from make_md indicating where the slide markdown source is found. By default, this function looks in the repo for the slides

output

An optional path indicating where the resulting slideshow should be written. Character scalar. If not provided, the slides will be saved into the lesson's git repository.

extra_flags

Character vector. Extra arguments to pass to pandoc to modify the conversion process

verbose

Logical scalar. TRUE if additional but non-essential logging should be provided.

open

Logical scalar. TRUE if you want the slides to be opened in your browser after they are generated.

Value

The path to the output slides, invisibly. Path to the reveal JS postprocessing lua filter